final styling for production
the production.css should be complete, the color errors have been taken care of. as for the production.pug i have also added buttons that appear showing the 'genre' and the 'network' the genre one redirects to the filtered home. the network one does nothing atm will work on it later. other than that it should be golden